feature: re-organise ExplosionSoundGroups
