+++ projects/pswg_gadgets/src/main/java/dev/pswg/item/grenades/soundGroups/ThermalDetonatorSoundGroup.java
@@ -1,4 +1,4 @@
-package dev.pswg.item;
+package dev.pswg.item.grenades.soundGroups;
 
 import dev.pswg.container.GadgetsSounds;
 import net.minecraft.world.entity.LivingEntity;
